cliplugins: move helpers into shared package
